Move siren types to appropriate place
types/ dir is reserved to declare types for other npm packages
